+++ src/main/java/com/intellij/plugins/haxe/model/HaxeMethodModel.java
@@ -236,7 +236,9 @@ public class HaxeMethodModel extends HaxeMemberModel implements HaxeExposableMod
   }
 
   public HaxeMethodModel getParentMethod(@Nullable HaxeGenericResolver resolver) {
-    final HaxeClassModel aClass = getDeclaringClass().getParentClass();
+    HaxeClassModel declaringClass = getDeclaringClass();
+    if(declaringClass == null) return null;
+    final HaxeClassModel aClass = declaringClass.getParentClass();
     return (aClass != null) ? aClass.getMethod(this.getName(), resolver) : null;
   }